refactors fixture list page components
